Rename render-lit-html.ts to render.ts and mark old exports as deprecated
