+++ packages/labs/ssr/src/index.ts
@@ -4,5 +4,5 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-export * from './lib/render-lit-html.js';
+export * from './lib/render.js';
 export {ElementRenderer} from './lib/element-renderer.js';
